Activate oscillator and show valve state on LED (#1)
Switch injector board to run off of the external 12MHz oscillator,
because it's faster and less jittery. Thus this updates both the
_XTAL_FREQ macro and the can_timing parameters so that the delay and CAN
functions both work without modification.

This commit also causes the injector_open function to illuminate the
white LED, and the injector_close function to deluminate it. This makes
it easier to debug when trying to determine if it's getting the goddamn
valve commands that I'm sending it.