Keep open/close functions from PWMing the mosfets
This fixes the situation in which, if injector open/close is called many
times in a row, the drive mosfets are essentially driven through PWM.
This is due to the 200us delay that occurs when switching valve position
in order to prevent shoot-through through the H bridge. New behaviour
is: if the valve is already in the desired state, do nothing and return.