Depower all valve drive pins at startup
When injector board boots up, as well as whenever the radio board loses
contact with RLCS, the injector board should stop applying power to all
valve driver pins. This is to prevent accidentally opening the valve
during fill, which would be very very bad.

Adds a special case whenever the requested valve state is VALVE_UNK to
turn off power to all of the valve pins. This will cause the limit
switches to both be depowered, which will cause the reported valve state
to be VALVE_UNK, so it all works out.