Make injector jog open valve for 200ms, then depower
This is empirically tested to work properly without nitrous, so here's
hoping that it works while the tank is full. This also checks for
repeated calls to injector_jog, two such calls in a row would fully open
the valve, which we don't want.

Also adds the case in the main loop to check if the radio board wants
the injector valve to be fully open.